fixed accidentally broken boolean extensions

+++ X10D.Performant/src/Custom/BooleanExtensions/BooleanExtensions.cs
@@ -14,7 +14,7 @@ namespace X10D.Performant.BooleanExtensions
         public static decimal ToDecimal(this bool value) => Unsafe.As<bool, byte>(ref value);
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToDouble"]'/>
-        public static double ToDouble(this bool value) => Unsafe.As<bool, long>(ref value);
+        public static double ToDouble(this bool value) => Unsafe.As<bool, int>(ref value);
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToInt16"]'/>
         public static short ToInt16(this bool value) => Unsafe.As<bool, short>(ref value);
@@ -23,13 +23,13 @@ namespace X10D.Performant.BooleanExtensions
         public static int ToInt32(this bool value) => Unsafe.As<bool, int>(ref value);
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToInt64"]'/>
-        public static long ToInt64(this bool value) => Unsafe.As<bool, long>(ref value);
+        public static long ToInt64(this bool value) => Unsafe.As<bool, int>(ref value);
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToNInt"]'/>
-        public static nint ToNInt(this bool value) => Unsafe.As<bool, nint>(ref value);
+        public static nint ToNInt(this bool value) => Unsafe.As<bool, int>(ref value);
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToUNInt"]'/>
-        public static nuint ToNUInt(this bool value) => Unsafe.As<bool, nuint>(ref value);
+        public static nuint ToNUInt(this bool value) => Unsafe.As<bool, uint>(ref value);
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToSByte"]'/>
         public static sbyte ToSByte(this bool value) => Unsafe.As<bool, sbyte>(ref value);
@@ -44,6 +44,6 @@ namespace X10D.Performant.BooleanExtensions
         public static uint ToUInt32(this bool value) => Unsafe.As<bool, uint>(ref value);
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToUInt64"]'/>
-        public static ulong ToUInt64(this bool value) => Unsafe.As<bool, ulong>(ref value);
+        public static ulong ToUInt64(this bool value) => Unsafe.As<bool, uint>(ref value);
     }
 }
